SL - fixed test file structure fixed getWorldPopulation methods and added testing

+++ src/main/java/com/napier/sem/Population.java
@@ -14,7 +14,7 @@ public class Population {
      * stores population
      */
 
-    public int population;
+    public long population;
 
 
 }
